Annotation.java
Here is the solution of the task. I was had to solve, truly complete the code(lines). The propose of this task is to train skills about annotations. It was little bit unfamiliar task. That's why I think I have to learn more about it. Actually every task of HackerRank is unfamiliar for me, however I will stay till the end.